Change behavior of `show` to issue `pause mouse close` + `quit`.
This fixes a number of issues. It makes the plots interactible by
default (#33) and it exposes an API to determine exactly when `gnuplot`
exits (#62).

The old behavior of `show` is now exposed as `show_and_keep_running`.

As explained in #33, we didn't do this earlier because it broke saving
with multiple formats and the animation example. By exposing both
behaviors, we serve both use uses.

Fixes #33
Fixes #62